adding gradient

+++ src/utils/render/render.h
@@ -6,7 +6,7 @@
 /*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 11:12:49 by mmarinel          #+#    #+#             */
-/*   Updated: 2022/11/05 12:19:12 by mmarinel         ###   ########.fr       */
+/*   Updated: 2022/11/05 12:55:24 by mmarinel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ t_bool	ft_iscolor(long long red, long long green, long long blue);
 void	ft_put_px_to_image(t_data *img_data, size_t px_offset, t_color color);
 size_t	ft_get_mlx_color(t_color color);
 t_bool	ft_fill_color(int red, int green, int blue, t_color *color);
-void	ft_apply_brightness(t_color *color, int factor);
+void	ft_apply_brightness(t_color *color, double factor);
 
 #endif
